+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changes
 
 - expand publisher filter to not push MergedPersons
+- quickfix publisher to always write an ndjson directly to S3
+  (will become outdated by MX-1808)
 
 ### Deprecated